Separate blockchain to its own file and improve the error message
