+++ apps/cargo-scout-audit/Cargo.toml
@@ -46,6 +46,8 @@ libloading = "0.8"
 regex = { version = "1.5", features = ["unicode"] }
 serde = { version = "1.0.163", features = ["derive"] }
 serde_json = "1.0"
+strum = "0.26"
+strum_macros = "0.26"
 tempfile = "3.8"
 tera = { version = "=1.20.0", features = ["builtins"] }
 toml = { version = "0.8.0" }
