+++ apps/cargo-scout-audit/src/detectors/builder.rs
@@ -6,8 +6,10 @@ use cargo_metadata::Metadata;
 use itertools::Itertools;
 
 use super::{configuration::DetectorConfiguration, library::Library, source::download_git_repo};
-use crate::startup::BlockChain;
-use crate::utils::{cargo_package, rustup};
+use crate::{
+    scout::blockchain::BlockChain,
+    utils::{cargo_package, rustup},
+};
 #[derive(Debug)]
 pub struct DetectorBuilder<'a> {
     cargo_config: &'a Config,
