+++ apps/cargo-scout-audit/src/detectors/configuration.rs
@@ -1,11 +1,12 @@
 use std::path::Path;
 
-use crate::startup::BlockChain;
 use anyhow::Result;
 use cargo::{
     core::{Dependency, GitReference, SourceId},
     util::IntoUrl,
 };
+
+use crate::scout::blockchain::BlockChain;
 
 #[derive(Debug, Clone)]
 pub struct DetectorConfiguration {
