+++ apps/cargo-scout-audit/src/lib.rs
@@ -2,5 +2,6 @@ extern crate lazy_static;
 
 pub mod detectors;
 pub mod output;
+pub mod scout;
 pub mod startup;
 pub mod utils;
