+++ apps/cargo-scout-audit/src/scout/mod.rs
@@ -1,0 +1,1 @@
+pub mod blockchain;
